feat: useInputMessage hook 구현
