+++ src/hooks/index.ts
@@ -1,2 +1,3 @@
 export { useAppDispatch } from './useAppDispatch';
 export { useAppSelector } from './useAppSelector';
+export { useInputMessage } from './useInputMessage';
